refactor names and interfaces
